Add structs, tuples, and pairs
The design we settled on is that these are effectively all the same type, but tuples and pairs have order and therefore also have the benefit of syntax. Structs are useful to represent largish abstract concepts that are not always backed by a JavaScript object, such as a URL record, request, or response. Tuples and pairs are useful for smaller, more fixed, abstract concepts.

Fixes #13, fixes #21, and fixes #81.