Change code unit-dependent algorithms to operate on JavaScript strings.
This change updates the code unit prefix and code unit less than
algorithms to operate on JavaScript strings, rather than strings, since
they only operate on them as sequences of code units. Additionally, the
code unit prefix algorithm mistakenly depends on the string's length in
a context when it needs to measure the number of code units, which this
change replaces to the JavaScript string's length.

Fixes #288.