Redo JavaScript and scalar value strings
This changes things a bit so there's one clear base type, called string. A scalar value string is now defined as a subset, similar to ASCII string.

A string has two different lengths, a code unit length and a code point length, and for now we keep the legacy length dfn (for JavaScript string and string) as we're not sure of all dependencies.

Fixes part of #291.