Redo JavaScript and scalar value strings
This changes things a bit so there's one clear base type, called string. A scalar value string is now defined as a subset of a string's value space, similar to ASCII string.

A string's length is in code units and a string also has a code point length for the couple places that need that.

Fixes part of #291. (The remainder there is fixing downstream dependencies.)