Fix and symmetrize "less than" and "prefix"
We had a definition of (code unit) "prefix" for strings, but "starts
with" for byte sequences, which led to usage errors within the
corresponding "less than" algorithms.

This makes "prefix" the primary operation for both strings and byte
sequences, with dedicated <dfn>s for "starts with" for both. Then, it
fixes "less than" for byte sequences to use "prefix" instead of "starts
with", which makes the algorithm correct and closes #309.

Additionally, this fixes the code unit prefix algorithm to be correct
instead of backward.