Change substring semantics to be about positions, not indices
Previously, the spec attempted to use code unit/point indices in the substring definitions. However, this mismatched how most programming languages do substring operations, and was buggy, as noted in #402.

This changes the definitions to operate on "positions" between code units/points, instead of indices, with a note explaining the difference. The resulting prose then matches most programming languages, as well as the JavaScript specification's less-formal "substring" definition.

Closes #402.